+++ src/index.ts
@@ -6,6 +6,7 @@ export * from './components/Checkbox';
 export * from './components/FormControl';
 export * from './components/Icon';
 export * from './components/IconButton';
+export * from './components/Avatar';
 export * from './components/Modal';
 export * from './components/Drawer';
 export * from './components/AppBar';